+++ src/test/java/cloud/filibuster/junit/tests/filibuster/smoke/local/hello/multiple/JUnitFilibusterHelloPartialHelloExternalGrpcTest.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.junit.tests.filibuster.smoke.local.hello;
+package cloud.filibuster.junit.tests.filibuster.smoke.local.hello.multiple;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
@@ -31,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  * Test simple annotation usage.
  */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFilibusterHelloPartialHelloExternalHttpWithSuppressCombinationsTest extends JUnitBaseTest {
+public class JUnitFilibusterHelloPartialHelloExternalGrpcTest extends JUnitBaseTest {
     private final static Set<String> testExceptionsThrown = new HashSet<>();
 
     private static int numberOfTestsExecuted = 0;
@@ -39,7 +39,7 @@ public class JUnitFilibusterHelloPartialHelloExternalHttpWithSuppressCombination
     private static int numberOfExceptionsThrown = 0;
 
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
-    @FilibusterTest(serverBackend=FilibusterLocalServerBackend.class, suppressCombinations=true, maxIterations=10)
+    @FilibusterTest(serverBackend=FilibusterLocalServerBackend.class, maxIterations=30)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
         ManagedChannel helloChannel = ManagedChannelBuilder
@@ -55,8 +55,8 @@ public class JUnitFilibusterHelloPartialHelloExternalHttpWithSuppressCombination
         Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian").build();
 
         try {
-            Hello.HelloReply reply = blockingStub.partialHelloExternalHttp(request);
-            assertEquals("Hello, Armerian World!!", reply.getMessage());
+            Hello.HelloReply reply = blockingStub.partialHelloExternalGrpc(request);
+            assertEquals("Hello, Hello, Hello, Armerian!!", reply.getMessage());
             assertFalse(wasFaultInjected());
         } catch (Throwable t) {
             numberOfExceptionsThrown++;
@@ -65,6 +65,7 @@ public class JUnitFilibusterHelloPartialHelloExternalHttpWithSuppressCombination
             boolean wasFaultInjected = wasFaultInjected();
 
             boolean firstRPCFailed = false;
+            boolean secondRPCFailed = false;
 
             if (wasFaultInjected) {
                 // First RPC failed.
@@ -104,27 +105,41 @@ public class JUnitFilibusterHelloPartialHelloExternalHttpWithSuppressCombination
 
                 // Second RPC failed.
 
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNKNOWN")) {
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DATA_LOSS: io.grpc.StatusRuntimeException: UNIMPLEMENTED")) {
                     expected = true;
+                    secondRPCFailed = true;
                 }
 
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: FAILED_PRECONDITION: HTTP RPC returned: 500")) {
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DATA_LOSS: io.grpc.StatusRuntimeException: INTERNAL")) {
                     expected = true;
+                    secondRPCFailed = true;
                 }
 
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: FAILED_PRECONDITION: HTTP RPC returned: 502")) {
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DATA_LOSS: io.grpc.StatusRuntimeException: UNAVAILABLE")) {
                     expected = true;
+                    secondRPCFailed = true;
                 }
 
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: FAILED_PRECONDITION: HTTP RPC returned: 503")) {
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DATA_LOSS: io.grpc.StatusRuntimeException: DEADLINE_EXCEEDED")) {
                     expected = true;
+                    secondRPCFailed = true;
+                }
+
+                if (secondRPCFailed) {
+                    boolean wasFaultInjectedOnWorldService = wasFaultInjectedOnService("HelloService");
+                    assertTrue(wasFaultInjectedOnWorldService);
+
+                    boolean wasFaultInjectedOnWorldMethod = wasFaultInjectedOnMethod("cloud.filibuster.examples.HelloService/Hello");
+                    assertTrue(wasFaultInjectedOnWorldMethod);
                 }
 
                 if (!expected) {
                     throw t;
                 }
             } else {
-                throw t;
+                if (!expected) {
+                    throw t;
+                }
             }
         }
 
@@ -143,13 +158,13 @@ public class JUnitFilibusterHelloPartialHelloExternalHttpWithSuppressCombination
     @Test
     @Order(3)
     public void testNumberOfTestsExecuted() {
-        assertEquals(9, numberOfTestsExecuted);
+        assertEquals(25, numberOfTestsExecuted);
     }
 
     @DisplayName("Verify correct number of exceptions thrown.")
     @Test
     @Order(4)
     public void numberOfExceptionsThrown() {
-        assertEquals(8, numberOfExceptionsThrown);
+        assertEquals(24, numberOfExceptionsThrown);
     }
 }
